requirements.txt
line-bot-sdk
pinecone
cohere
groq
requests
urllib3<2.0.0
